added unit tests to pipeline
